Move Layout to HOC folder
