+++ web-app/src/hoc/index.js
@@ -1,0 +1,2 @@
+export { default as withAuth } from './withAuth';
+export { default as Layout } from './Layout';
